Arrumando um pouco para o padrão de projeto

+++ src/controllers/PacienteDao.java
@@ -3,30 +3,34 @@ package src.controllers;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import src.data.Banco;
-import src.models.FuncUtils;
 import src.models.Paciente;
+import src.utils.FuncUtils;
+
 import java.sql.Date;
 
 public class PacienteDao {
+    static Banco db;
 
-    public static void cadastrarPaciente(Banco db) {
-        FuncUtils.clearScreen();
-        Paciente paciente = new Paciente();
-        paciente.preenchePaciente();
+    public PacienteDao(){
+        db = new Banco();
+    }
+
+    public static void cadastrarPaciente(Paciente paciente) {
+      
         String querry = String.format(
                 "INSERT INTO Paciente (nome, cpf, telefone, data_nascimento, sexo, internado, plano_saude)VALUES ('%s', '%s', '%s', '%tF', %b, %b, %b);",
                 paciente.getNome(), paciente.getCpf(), paciente.getTelefone(), paciente.getDataNasc(),
                 paciente.getSexo(),
                 paciente.isInternado(), paciente.isPlanoDeSaude());
-        db.querry_insup(querry);
+                db.querry_insup(querry);
     }
 
-    public static Paciente buscaPaciente(String codPaciente, Banco db) throws SQLException {
+    public static Paciente buscaPaciente(String codPaciente) throws SQLException {
         String query = "SELECT * FROM Paciente WHERE id_paciente = '" + codPaciente + "';";
         ResultSet rs = db.querry_busca(query);
         Paciente paciente = new Paciente();
 
-        if (rs.next()) {
+        if (rs.next() && rs != null) {
             paciente.setNome(rs.getString("nome"));
             paciente.setCpf(rs.getString("cpf"));
             paciente.setTelefone(rs.getString("telefone"));
@@ -40,44 +44,36 @@ public class PacienteDao {
             paciente.setPlanoDeSaude(rs.getBoolean("plano_saude"));
             paciente.setInternado(rs.getBoolean("internado"));
         }
-        if (paciente.getNome().length() < 3) {
+        else{
             System.out.println("Paciente não encontrado.");
             return null;
         }
         return paciente;
     }
 
-    public static void editarPaciente(Banco db) throws SQLException {
-        listarPacientes(db);
-        String codPaciente = FuncUtils.readCod();
-        Paciente paciente = buscaPaciente(codPaciente, db);
-        if (paciente.getNome().length() < 3) {
-            System.out.println("Paciente não encontrado.");
-            return;
-        }
-
-        FuncUtils.clearScreen();
-        if (!paciente.editaPaciente())
-            return;
-        String querry = String.format(
+    public static void editarPaciente(String codPaciente) throws SQLException {
+        Paciente paciente = buscaPaciente(codPaciente);
+        if (paciente!= null) {
+            String querry = String.format(
                 "UPDATE Paciente SET nome = '%s', cpf = '%s', telefone = '%s', data_nascimento = '%tF', sexo = %b, internado = %b, plano_saude = %b WHERE id_paciente = '%s';",
                 paciente.getNome(), paciente.getCpf(), paciente.getTelefone(), paciente.getDataNasc(),
                 paciente.getSexo(),
                 paciente.isInternado(), paciente.isPlanoDeSaude(), paciente.getCodPaciente());
-        db.querry_insup(querry);
+                db.querry_insup(querry);
+        } 
+        
     }
 
-    public static void excluirPaciente(Banco db) throws SQLException {
+    public static void excluirPaciente(String codPaciente) throws SQLException {
         FuncUtils.clearScreen();
-        listarPacientes(db);
-
-        String codPaciente = FuncUtils.readCod();
-
-        String querry = "DELETE FROM Paciente WHERE id_paciente = '" + codPaciente + "';";
-        db.querry_insup(querry);
+        Paciente control = buscaPaciente(codPaciente);
+        if (control != null) {
+            String querry = "DELETE FROM Paciente WHERE id_paciente = '" + codPaciente + "';";
+            db.querry_insup(querry); 
+        }
     }
 
-    public static void listarPacientes(Banco db) throws SQLException {
+    public static void listarPacientes() throws SQLException {
         String query = "SELECT * FROM Paciente;";
         ResultSet rs = db.querry_busca(query);
 
@@ -106,25 +102,20 @@ public class PacienteDao {
         }
     }
 
-    public static void internar(Banco db) throws SQLException {
-        listarPacientes(db);
-        String codPaciente = FuncUtils.readCod();
+    public static void internar(String codPaciente) throws SQLException {
 
-        Paciente aux = buscaPaciente(codPaciente, db);
-        if (aux.getNome().length() < 3) {
-            System.out.println("Paciente não encontrado.");
-            return;
+        Paciente aux = buscaPaciente(codPaciente);
+        if (aux != null) {
+            String querry = "UPDATE Paciente SET internado = 1 WHERE id_paciente = '" + codPaciente + "';";
+            db.querry_insup(querry);
+             System.out.println("\n" + aux.getNome() + " foi internado.");
         }
-        String querry = "UPDATE Paciente SET internado = 1 WHERE id_paciente = '" + codPaciente + "';";
-        db.querry_insup(querry);
-        System.out.println("\n" + aux.getNome() + " foi internado.");
+        
     }
 
-    public static void alta(Banco db) throws SQLException {
-        listarPacientes(db);
-        String codPaciente = FuncUtils.readCod();
-
-        Paciente aux = buscaPaciente(codPaciente, db);
+    public static void alta(String codPaciente) throws SQLException {
+        listarPacientes();
+        Paciente aux = buscaPaciente(codPaciente);
         if (aux.getNome().length() < 3) {
             System.out.println("Paciente não encontrado.");
             return;
